+++ .idea/modules.xml
@@ -5,7 +5,6 @@
       <module fileurl="file://$PROJECT_DIR$/UFU.iml" filepath="$PROJECT_DIR$/UFU.iml" />
       <module fileurl="file://$PROJECT_DIR$/UFU_2017.iml" filepath="$PROJECT_DIR$/UFU_2017.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
-      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
 </project>
