+++ app/src/main/java/wildbakery/ufu/Adapters/ItemsAdapterEvent.java
@@ -1,4 +1,4 @@
-package wildbakery.ufu.Adapter;
+package wildbakery.ufu.Adapters;
 
 /**
  * Created by DIKII PEKAR on 13.02.2017.
@@ -16,7 +16,7 @@ import com.squareup.picasso.Picasso;
 import java.util.List;
 
 import wildbakery.ufu.Constants;
-import wildbakery.ufu.Model.EventItem;
+import wildbakery.ufu.Models.EventItem;
 import wildbakery.ufu.R;
