+++ app/src/main/java/wildbakery/ufu/Adapters/ItemsAdapterSale.java
@@ -1,4 +1,4 @@
-package wildbakery.ufu.Adapter;
+package wildbakery.ufu.Adapters;
 
 /**
  * Created by DIKII PEKAR on 19.12.2016.
@@ -18,7 +18,7 @@ import com.squareup.picasso.Picasso;
 import java.util.List;
 
 import wildbakery.ufu.Constants;
-import wildbakery.ufu.Model.SaleItem;
+import wildbakery.ufu.Models.SaleItem;
 import wildbakery.ufu.R;
 
 public class ItemsAdapterSale extends  RecyclerView.Adapter<ItemsAdapterSale.ViewHolder>  {
