+++ app/src/main/java/wildbakery/ufu/FetchDataPackage/VuzAPI.java
@@ -1,13 +1,14 @@
-package wildbakery.ufu.Interfaces;
+package wildbakery.ufu.FetchDataPackage;
 
 import retrofit2.Call;
 import retrofit2.Retrofit;
 import retrofit2.converter.gson.GsonConverterFactory;
 import retrofit2.http.GET;
-import wildbakery.ufu.Model.Event.EventModel;
-import wildbakery.ufu.Model.Job.JobsModel;
-import wildbakery.ufu.Model.News.NewsModel;
-import wildbakery.ufu.Model.Sale.SaleModel;
+import wildbakery.ufu.Models.EventItem;
+import wildbakery.ufu.Models.JobItem;
+import wildbakery.ufu.Models.NewsItem;
+import wildbakery.ufu.Models.QueryModel;
+import wildbakery.ufu.Models.SaleItem;
 
 import static wildbakery.ufu.Constants.HTTP.BASE_URL;
 
@@ -16,28 +17,28 @@ import static wildbakery.ufu.Constants.HTTP.BASE_URL;
  * Created by DIKII PEKAR on 01.12.2016.
  */
 
-public interface APIservice {
+public interface VuzAPI {
 
     @GET("v1/jobs.json")
-    Call<JobsModel> getAllJob();
+    Call<QueryModel<JobItem>> getJobs();
 
     @GET("v1/news.json")
-    Call<NewsModel> getAllNews();
+    Call<QueryModel<NewsItem>> getNews();
 
     @GET("v1/events.json")
-    Call<EventModel> getAllEvent();
+    Call<QueryModel<EventItem>> getEvents();
 
     @GET("v1/sales.json")
-    Call<SaleModel> getAllSale();
+    Call<QueryModel<SaleItem>> getSales();
 
     class Factory {
-        private  static  APIservice service;
-        public static  APIservice getInstance(){
+        private  static VuzAPI service;
+        public static VuzAPI getInstance(){
             if (service == null){
                 Retrofit retrofit = new Retrofit.Builder().baseUrl(BASE_URL)
                         .addConverterFactory(GsonConverterFactory.create())
                         .build();
-                service = retrofit.create(APIservice.class);
+                service = retrofit.create(VuzAPI.class);
                 return service;
             }
             else {
