+++ app/src/main/java/wildbakery/ufu/Fragments/FragmentNews.java
@@ -1,4 +1,4 @@
-package wildbakery.ufu.Fragment;
+package wildbakery.ufu.Fragments;
 
 
 import android.os.Bundle;
@@ -17,25 +17,24 @@ import java.util.List;
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
-import wildbakery.ufu.Adapter.ItemsAdapterNews;
-import wildbakery.ufu.Interfaces.APIservice;
-import wildbakery.ufu.Model.NewsItem;
-import wildbakery.ufu.Model.QueryModel;
+import wildbakery.ufu.Adapters.ItemsAdapterNews;
+import wildbakery.ufu.FetchDataPackage.VuzAPI;
+import wildbakery.ufu.Models.NewsItem;
+import wildbakery.ufu.Models.QueryModel;
 import wildbakery.ufu.R;
 
 import static wildbakery.ufu.R.id.recyclerviewNews;
+import static wildbakery.ufu.R.id.submenuarrow;
 
 
 /**
  * A simple {@link Fragment} subclass.
  */
-public class FragmentNews extends BaseFragment {
+public class FragmentNews extends FragmentPage implements ItemsAdapterNews.OnItemClickListener{
     private static final String TAG = "FragmentNews";
 
-    private RecyclerView recyclerView;
     private List<NewsItem> listItems;
     private ItemsAdapterNews mAdapter;
-    private LinearLayoutManager mLayoutManager;
     private DetailFragmentNews activeDetailFragment;
 
     public FragmentNews() {
@@ -52,29 +51,7 @@ public class FragmentNews extends BaseFragment {
                              Bundle savedInstanceState) {
 
 
-        View view = inflater.inflate(R.layout.fragment_news, container, false);
-        recyclerView = (RecyclerView) view.findViewById(recyclerviewNews);
-
-
-        Log.v(TAG, "onCreateView()");
-        APIservice.Factory.getInstance().getAllNews().enqueue(new Callback<QueryModel<NewsItem>>() {
-
-            @Override
-            public void onResponse(Call<QueryModel<NewsItem>> call, Response<QueryModel<NewsItem>> response) {
-                if (response.isSuccess()) {
-                    Log.v(TAG, "refresh");
-                    listItems = new ArrayList<>();
-                    QueryModel<NewsItem> result = response.body();
-                    listItems = result.getItems();
-                    setRecyclerView();
-                }
-            }
-
-            @Override
-            public void onFailure(Call<QueryModel<NewsItem>> call, Throwable t) {
-
-            }
-        });
+        View view = super.onCreateView(inflater, container, savedInstanceState);
 
 
         return view;
@@ -98,11 +75,6 @@ public class FragmentNews extends BaseFragment {
         mAdapter = new ItemsAdapterNews(listItems, new ItemsAdapterNews.OnItemClickListener() {
             @Override
             public void onItemClick(NewsItem item) {
-                Log.d(getClass().getCanonicalName(), "onItemClick: item = " + item);
-                activeDetailFragment = DetailFragmentNews.newInstance(item);
-
-                getActivity().getSupportFragmentManager().beginTransaction().replace(R.id.viewNews, activeDetailFragment).commit();
-                recyclerView.setVisibility(View.GONE);
             }
         });
 
@@ -112,5 +84,13 @@ public class FragmentNews extends BaseFragment {
         recyclerView.setLayoutManager(mLayoutManager);
         recyclerView.setItemAnimator(new DefaultItemAnimator());
         recyclerView.setAdapter(mAdapter);
+    }
+
+    @Override
+    public void onItemClick(NewsItem item) {
+        Log.d(getClass().getCanonicalName(), "onItemClick: item = " + item);
+        activeDetailFragment = DetailFragmentNews.newInstance(item);
+        getActivity().getSupportFragmentManager().beginTransaction().replace(R.id.viewNews, activeDetailFragment).commit();
+        recyclerView.setVisibility(View.GONE);
     }
 }
