+++ app/src/main/java/wildbakery/ufu/Models/JobItem.java
@@ -1,12 +1,10 @@
 
-package wildbakery.ufu.Model;
+package wildbakery.ufu.Models;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
 import java.io.Serializable;
-
-import wildbakery.ufu.Model.Item;
 
 public class JobItem extends Item implements Serializable {
