+++ app/src/main/java/wildbakery/ufu/Models/QueryModel.java
@@ -1,4 +1,4 @@
-package wildbakery.ufu.Model;
+package wildbakery.ufu.Models;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
