+++ app/src/main/res/layout/fragment_event.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="wildbakery.ufu.Fragment.FragmentEvent">
+    tools:context="wildbakery.ufu.Fragments.FragmentEvent">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/recyclerviewEvent"
