+++ app/src/main/res/layout/fragment_news.xml
@@ -4,7 +4,7 @@
               android:layout_width="match_parent"
               android:layout_height="match_parent"
               android:orientation="vertical"
-              tools:context="wildbakery.ufu.Fragment.FragmentJob">
+              tools:context="wildbakery.ufu.Fragments.FragmentJob">
 
 
     <android.support.v7.widget.RecyclerView
